+++ CHANGELOG.md
@@ -11,4 +11,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Issue-6: Confirm panel to reset project
+- Issue #5: Import file
+- Issue #6: Confirm panel to reset project